+++ domino_data/logging.py
@@ -1,6 +1,7 @@
 """Logging module."""
 
 import os
+import sys
 import tempfile
 
 from loguru import logger as _logger
@@ -24,6 +25,10 @@ def getlogger():
                 "retention": 7,
                 "enqueue": True,
                 "serialize": True,
+            },
+            {
+                "format": "[{time}] {message}",
+                "sink": sys.stdout,
             },
         ],
         extra={
